+++ client/src/components/User/Post.tsx
@@ -2,16 +2,15 @@ import { useState, useRef } from "react";
 import { useParams, useNavigate, useLocation } from "react-router-dom";
 
 import { axios } from "../../config";
-import VerifiedIcon from "../../assets/verified-check.png";
 
 import type { Post } from "../../context/PostContext";
 import { usePosts } from "../../context/PostContext";
 import { useUser } from "../../context/UserContext";
 
-import { GetTimePast, FormatNumber } from "../../utilities/Format";
 import SwalUtility from "../../utilities/SwalUtility";
 
 import Comments from "./Helper/Posts/Comments";
+import PostCardLayout from "./Helper/Posts/PostCardLayout";
 
 export default function Post({ modal = false }) {
     const { ownerId, id } = useParams();
@@ -32,6 +31,8 @@ export default function Post({ modal = false }) {
             navigate("/");
         }
     }
+
+    const isSelf = user?.id === ownerId;
 
     const HandleReply = (commentOwner: string, commentId: number) => {
         setComment(`@${commentOwner} ${comment}`);
@@ -61,18 +62,6 @@ export default function Post({ modal = false }) {
         }
     }
 
-    const HandleDeletion = async (postId: number, postOwner: number, fn: () => void) => {
-        let security_code: string = "";
-
-        if (user?.role_id! >= 2 && postOwner !== user?.id) {
-            security_code = (await SwalUtility.AskSecurityCode(user?.code))!;
-            if (!security_code)
-                return;
-        }
-
-        await DeleteItem(postId, fn, security_code);
-    }
-
     const post: Post | undefined = posts.find(p => p.postId === +id! && p.ownerId === +ownerId!);
 
     if (loading || !post || !user)
@@ -81,8 +70,8 @@ export default function Post({ modal = false }) {
     return (
         <div style={{
             position: modal ? "fixed" : "relative", top: modal ? "50%" : "0", left: modal ? "50%" : "0",
-            transform: modal ? "translate(-50%, -50%)" : "none", background: "white", padding: "20px", borderRadius: modal ? "12px" : "0",
-            width: modal ? "800px" : "100%", height: modal ? "600px" : "100%", boxShadow: modal ? "0 0 20px rgba(0,0,0,0.3)" : "none",
+            transform: modal ? "translate(-50%, -50%)" : "none", background: "white", padding: "10px", borderRadius: modal ? "12px" : "0",
+            height: modal ? "90vh" : "100%", width: modal ? "95%" : "100%", maxWidth: modal ? "800px" : "100%", maxHeight: modal ? "600px" : "100%", boxShadow: modal ? "0 0 20px rgba(0,0,0,0.3)" : "none",
             overflowY: modal ? "auto" : "visible", zIndex: modal ? 1000 : "auto", overflowX: "hidden", display: "flex", flexDirection: "column"
         }}>
             <div className="modal-header justify-content-center position-relative">
@@ -94,67 +83,7 @@ export default function Post({ modal = false }) {
             <hr />
             <div className="modal-body flex-grow-1" style={{ overflowY: 'auto', paddingBottom: '80px' }}>
                 {/* Post side */}
-                <div className="card rounded-0 p-0">
-                    <div className="card-body mt-1 border-bottom-0 rounded-4 bg-white">
-                        <div className="d-flex justify-content-between">
-                            <div className="d-flex align-items-center text-break text-wrap" onClick={() => navigate(`/profile/${post.username}`)}>
-                                <img src={`/public/profile-pics/${post.profileUrl}`} alt="" className="rounded-circle" style={{ height: '50px', width: '50px', minHeight: '50px', minWidth: '50px', objectFit: 'cover' }} />
-                                <div className="ps-3">
-                                    <div className="d-flex align-items-center">
-                                        <span className={"fw-semibold fs-5"}>{post.full_name}</span>
-                                        {
-                                            post.verified && (<img src={VerifiedIcon} alt="" className="rounded-circle pt-1" style={{ paddingLeft: '4px', height: '25px' }} />)
-                                        }
-                                    </div>
-                                    <div className="text-muted fs-6">
-                                        <span style={{ color: post.role_id >= 2 ? 'goldenrod' : 'inherit', fontWeight: '500' }}>(@{post.username})</span>
-                                    </div>
-                                </div>
-                            </div>
-                            {post.createdAt && (<span className="text-muted fs-6">{GetTimePast(post.createdAt)}</span>)}
-                        </div>
-                        <div className="post-body">
-                            <div className="d-flex my-4 text-break text-wrap" style={{ paddingLeft: 'calc(115px - 50px)' }}>
-                                {post.description}
-                            </div>
-                            {
-                                post.imageUrl && (
-                                    <img src={`/public/${post.imageUrl}`} alt="" className="rounded-0"
-                                        style={{
-                                            marginLeft: 'calc(115px - 50px)', marginRight: 'calc(115px - 50px)',
-                                            height: '350px', width: '350px', objectFit: 'cover'
-                                        }} />
-                                )
-                            }
-                        </div>
-                    </div>
-                    <div className="card-footer bg-white">
-                        { /* Like, comment, etc box here. (Imported from PostCard.tsx) */}
-                        <div className="d-flex align-items-center" style={{ paddingLeft: '65px' }}>
-                            <div>
-                                <i className={`fas fa-heart text-${post.liked ? "danger" : "white"}`}
-                                    style={{ WebkitTextStroke: post.liked ? '' : '1px black' }} onClick={() => AddLike(post.postId)}></i>
-                                <span className="px-1">{FormatNumber(post.likeCount)}</span>
-                            </div>
-                            <div className="ms-3">
-                                <i className="fas fa-comment text-white" style={{ WebkitTextStroke: '1px black' }}></i>
-                                <span className="px-1">{FormatNumber(post.commentCount)}</span>
-                            </div>
-                            {user.id !== post.ownerId && (
-                                <div className="ms-3">
-                                    <i className="fas fa-triangle-exclamation" onClick={() => ReportPost(post.postId)}></i>
-                                    <span className="px-1">Report</span>
-                                </div>
-                            )}
-                            {(user.role_id >= 2 || user.id === post.ownerId) && (
-                                <div className="ms-3 text-danger" onClick={() => HandleDeletion(post?.postId, post?.ownerId, CloseModal)}>
-                                    <i className="fas fa-trash"></i>
-                                    <span className="px-1">Delete</span>
-                                </div>
-                            )}
-                        </div>
-                    </div>
-                </div>
+                <PostCardLayout post={post} user={user} isMain={true} isSelf={isSelf} actions={{ AddLike, DeleteItem, ReportPost }} />
                 <hr className="my-5" />
                 {/* Comments section here */}
                 {